Fixed major bug where, when appending data to the URL, it stored a signed int in a size_t and ended up with a bunch of invalid URL elements. Added "Errors" response which shows up when there is a URL error. Fixed bugs where context was actually user data. Changed start/end element to use full namespace-equiped ones (SAX2). Modified createHashtableFromAtts so that it uses the new attributes from SAX2. Parsing now uses qualified names.

+++ BBX/src/bing.c
@@ -512,7 +512,7 @@ void append_data(hashtable_t* table, const char* format, const char* key, void**
 	char buffer[256];
 	buffer[0] = '\0';
 	char* rett;
-	size_t size = hashtable_get_item(table, key, NULL);
+	int size = hashtable_get_item(table, key, NULL);
 	if(size > 0)
 	{
 		//Make sure that the data buffer (for getting the data) is large enough (add one for safety)
